requirements.txt
# Agents
langchain==0.3.14
langchain-community==0.3.14
python-dotenv==1.0.1
wikipedia==1.4.0
duckduckgo-search==7.2.1
openai==1.59.6

# Assemblers

# Classifiers
torch
torchvision
pytorch-lightning
scikit-learn
matplotlib
Pillow
numpy
opencv-python
kagglehub
selenium
requests
beautifulsoup4
lxml

# Diffusion Model
diffusers
transformers
torch
torchvision
Pillow
accelerate
protobuf
sentencepiece

#main
opencv-python
PyQt5

